feat: auto fetch latest articles in news API
